Replace ViewController+SwinjectSpec with ViewController+SwinjectTests.
